pyproject.toml
[project]
name = "earthcode"
version = "0.1.7"
description = "Tools for creating, validating, and searching EarthCODE Open Science Catalog metadata."
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    { name = "Krasen Samardzhiev", email = "krasensam@gmail.com" },
    { name = "Deyan Samardzhiev", email = "dean@lampata.co.uk" },
    { name = "Claudio Candelori", email = "claudio.candelori@cgi.com"},
]
keywords = [
    "earth observation",
    "earthcode",
    "geospatial"
]

classifiers = [
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "pystac>=1.14.1,<2",
    "xarray>=2025.12.0,<2026",
    "shapely>=2.1.2,<3",
    "xstac>=1.2.0,<2",
    "zarr>=3.1.5,<4",
    "dask>=2025.12.0,<2026",
    "netcdf4>=1.7.3,<2",
    "bottleneck>=1.6.0,<2",
    "rioxarray>=0.20.0,<0.21",
    "geopandas>=1.1.2,<2",
    "pyarrow>=23.0.0,<24",
    "pylance>=2.0.1,<3",
    "pillow>=10.3,<12",
    "fsspec>=2025.10.0,<2026",
    "jsonschema>=4.25.1,<5",
    "numpy>=2.4.1,<3",
    "onnxruntime>=1.23.2,<2",
    "pandas>=3.0.0,<4",
    "platformdirs>=4.5.0,<5",
    "pydantic>=2.12.5,<3",
    "requests>=2.32.5,<3",
    "tokenizers>=0.22.2,<1",
    "PyYAML"
]

[project.urls]
Homepage = "https://github.com/ESA-EarthCODE/earthcode-library"
Repository = "https://github.com/ESA-EarthCODE/earthcode-library"
Issues = "https://github.com/ESA-EarthCODE/earthcode-library/issues"
Documentation = "https://esa-earthcode.github.io/earthcode-library/"

[project.optional-dependencies]
dev = [
    "jupyter-book>=1.0.4,<2",
    "jupyterlab>=4.5.0,<5",
    "papermill>=2.7.0,<3",
    "pytest>=9.0.2,<10",
    "pytest-cov>=7.0.0,<8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
exclude = ["earthcode/models/**"]

[tool.hatch.build.targets.wheel]
packages = ["earthcode"]

[project.scripts]
earthcode_template_gen = "earthcode.template_generator:main"
earthcode_stac_gen = "earthcode.stac_generator:main"